# Improve JSON.generate() performance