This pull request adds some shortcut to convert Object to json.
JSON.generate() performance will be 3.53 times faster by this.

## Environment
* Machine : MacBook (Retina, 12-inch, 2017)
* OS : macOS 10.13.3
* Compiler : Apple LLVM version 9.1.0 (clang-902.0.31)
* Ruby : ruby 2.5.0p0 (2017-12-25 revision 61468) [x86_64-darwin17]

## Before
```
$ ruby bench_json_generate2.rb
Warming up --------------------------------------
                json    54.000  i/100ms
                yajl    63.000  i/100ms
                  oj   286.000  i/100ms
Calculating -------------------------------------
                json    552.420  (± 1.3%) i/s -      2.808k in   5.083937s
                yajl    640.489  (± 1.7%) i/s -      3.213k in   5.017937s
                  oj      2.961k (± 4.1%) i/s -     14.872k in   5.032250s
```

## After
```
$ ruby bench_json_generate2.rb
Warming up --------------------------------------
                json   176.000  i/100ms
                yajl    50.000  i/100ms
                  oj   234.000  i/100ms
Calculating -------------------------------------
                json      1.949k (± 4.3%) i/s -      9.856k in   5.068917s
                yajl    643.232  (± 1.4%) i/s -      3.250k in   5.053518s
                  oj      2.970k (± 4.2%) i/s -     14.976k in   5.053368s
```

![graph](https://user-images.githubusercontent.com/199156/37039999-2cdf69f6-219c-11e8-9b5b-cd78bf93407a.png)

## Benchmark code
```ruby
require 'json'
require 'oj'
require 'yajl'
require 'benchmark/ips'

obj = []

1000.times do |i|
  obj << {
    "id" => i,
    :age => 42,
  }
end

Benchmark.ips do |x|
  x.report("json") { JSON.generate(obj) }
  x.report("yajl") { Yajl::Encoder.encode(obj, nil) }
  x.report("oj")   { Oj.dump(obj) }
end
```