@searls 2018-06-01
This seems like a pretty impressive improvement for such a small diff! Good work!