@glebm 2019-04-21
@flori Anything preventing this from getting merged? It gives a nice 3.5x performance increase for `JSON.generate`.