@Watson1978 2019-05-12
Seems my patches was merged into master https://github.com/flori/json/compare/167ada8da7299fd27314fadad3797b48cee442b1...91a24ecac3d722bc3e545fdd0657ed225e91c2df

Thanks.